feat: Splitted loginController from UsersController
